Bow has Charge Levels
-Dodging is tied to left stick input
-There is a bool on PlayerControlelr to toggle the AddForce dodge, check it out
-Now there are multiple charge levels
-Holding charge will change player color and increase the damage and speed of the arrows
-Jump is now left trigger and A button
-Dodge is now left bumper and B button. This is so we can test things easier.

+++ Assets/Scripts/Projectile.cs
@@ -9,7 +9,7 @@ public class Projectile : MonoBehaviour { //should this just be named arrow?
 
     public Rigidbody rb;
 
-    public int speed = 1;
+    public float speed = 1f;
     public int damage = 100;
 
     public float lifetime = 3f;
